Workin' :) TODO : test it all, and give daily prizes

+++ Server/Stump.Server.WorldServer/Game/Actors/Fight/SummonedMonster.cs
@@ -78,7 +78,7 @@ namespace Stump.Server.WorldServer.Game.Actors.Fight
                 Id,
                 Look.GetEntityLook(),
                 GetEntityDispositionInformations(client),
-                Team.Id,
+                (sbyte)Team.Id,
                 IsAlive(),
                 GetGameFightMinimalStats(client),
                 (short)Monster.Template.Id,
